move files into folders, change class declarations to interfaces, use map to ensure bar service repsonse is enforced to type constraint
