+++ src/app/bar-service/bar.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/internal/Observable';
-import { Bar } from './bar';
+import { Bar, BarServiceResponse } from '../common/interfaces/bar';
+import { map } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
@@ -11,6 +12,7 @@ export class BarService {
   constructor(private http: HttpClient) {}
 
   getAllBars(): Observable<Bar[]> {
-    return this.http.get<Bar[]>('http://localhost:64001/barinfo');
+    return this.http.get<BarServiceResponse>('http://localhost:64001/barinfo')
+      .pipe(map((response: BarServiceResponse) => response.bars));
   }
 }
