--- src/app/bar.ts
@@ -1,4 +1,0 @@
-export class Bar {
-    id: number;
-    name: string;
-}
