+++ src/app/common/interfaces/bar.ts
@@ -1,0 +1,8 @@
+export interface Bar {
+    id: number;
+    name: string;
+}
+
+export interface BarServiceResponse {
+    bars: Bar [];
+}
