+++ src/app/common/interfaces/drink.ts
@@ -1,4 +1,4 @@
-export class Drink {
+export interface Drink {
     drink_id: number;
     bar_id: number;
     name: string;
